`helio-core`新增内置枚举`EnabledStatusEnum`、`GenderEnum`、`YesOrNoEnum` pref: 代码风格调整 docs: 补充HelioBaseEntity.delFlag字段注释
