+++ helio-core/src/main/java/cc/uncarbon/framework/core/enums/GenderEnum.java
@@ -1,0 +1,24 @@
+package cc.uncarbon.framework.core.enums;
+
+import com.baomidou.mybatisplus.annotation.EnumValue;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+
+/**
+ * 生理性别枚举类
+ */
+@AllArgsConstructor
+@Getter
+public enum GenderEnum implements HelioBaseEnum<Integer> {
+
+    UNKNOWN(0, "未知"),
+    MALE(1, "男"),
+    FEMALE(2, "女"),
+
+    ;
+    @EnumValue
+    private final Integer value;
+    private final String label;
+
+}
