+++ helio-core/src/main/java/cc/uncarbon/framework/core/enums/YesOrNoEnum.java
@@ -1,20 +1,18 @@
-package cc.uncarbon.framework.crud.enums;
+package cc.uncarbon.framework.core.enums;
 
-import cc.uncarbon.framework.core.enums.HelioBaseEnum;
 import com.baomidou.mybatisplus.annotation.EnumValue;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-
 /**
- * 启用状态枚举
+ * 是或否枚举
  */
 @AllArgsConstructor
 @Getter
-public enum EnabledStatusEnum implements HelioBaseEnum<Integer> {
+public enum YesOrNoEnum implements HelioBaseEnum<Integer> {
 
-    DISABLED(0, "禁用"),
-    ENABLED(1, "启用"),
+    NO(0, "否"),
+    YES(1, "是"),
 
     ;
     @EnumValue
@@ -26,17 +24,17 @@ public enum EnabledStatusEnum implements HelioBaseEnum<Integer> {
      * @param value 外部值
      * @return null or 枚举对象
      */
-    public static EnabledStatusEnum of(Integer value) {
+    public static YesOrNoEnum of(Integer value) {
         if (value == null) {
             return null;
         }
 
         // 区分度小，直接用if判断了
-        if (DISABLED.value.equals(value)) {
-            return DISABLED;
+        if (YES.value.equals(value)) {
+            return YES;
         }
-        if (ENABLED.value.equals(value)) {
-            return ENABLED;
+        if (NO.value.equals(value)) {
+            return NO;
         }
         return null;
     }
@@ -46,16 +44,16 @@ public enum EnabledStatusEnum implements HelioBaseEnum<Integer> {
      * @param old 原值
      * @return 新值
      */
-    public static EnabledStatusEnum reverse(EnabledStatusEnum old) {
+    public static YesOrNoEnum reverse(YesOrNoEnum old) {
         if (old == null) {
             return null;
         }
 
         switch (old) {
-            case DISABLED:
-                return ENABLED;
-            case ENABLED:
-                return DISABLED;
+            case YES:
+                return NO;
+            case NO:
+                return YES;
         }
 
         throw new IllegalArgumentException();
